update split minichat in component
