+++ srcs/django_code/transcend/myapp/static/html/components/minichat/message.js
@@ -1,0 +1,34 @@
+import { Component } from "../../../js/component.js";
+
+export class Message extends Component
+{
+
+    static getName()
+    {
+        return "message"
+    }
+
+    rules()
+    {
+        return {
+            content: ['required'],
+            who: ['required']
+        }
+    }
+
+    connectedCallback()
+    {
+        super.connectedCallback()
+
+        const who = this.who == "reciever" ? "recieveMsg" : "sendMsg"
+
+        this.innerHTML = `
+        <div class="msgContainer">
+            <div class="${who}">
+                <p>${this.content}</p>
+            </div>
+        </div>
+        `
+    }
+
+}
